+++ examples/specific.py
@@ -1,0 +1,9 @@
+x = 1
+y = 0
+
+try:
+    x / y
+except Exception:
+    import debuglater
+    debuglater.run(__file__)
+    raise