+++ src/debuglater/__init__.py
@@ -1,6 +1,6 @@
-from debuglater.pydump import run
+from debuglater.pydump import run, excepthook_factory
 from debuglater.ipython import patch_ipython
 
-__all__ = ['run', 'patch_ipython']
+__all__ = ['run', 'excepthook_factory', 'patch_ipython']
 
 __version__ = "1.4.2dev"