+++ src/debuglater/pydump.py
@@ -28,6 +28,7 @@ import sys
 import pdb
 import gzip
 import linecache
+from traceback import format_exception
 try:
     import cPickle as pickle
 except ImportError:
@@ -293,15 +294,26 @@ def _cache_files(files):
         linecache.cache[name] = (len(data), None, lines, name)
 
 
-def run(filename, echo=True):
+def run(filename, echo=True, tb=None):
     out = Path(filename).with_suffix('.dump')
 
     if echo:
         print(Fore.RED + f'Exception caught, writing {out}\n')
-    save_dump(out)
+
+    save_dump(out, tb=tb)
+
     if echo:
         print(f'To debug, run:\n  dltr {out}')
         print(Style.RESET_ALL)
+
+
+def excepthook_factory(filename):
+
+    def excepthook(type, value, traceback):
+        print(''.join(format_exception(type, value, traceback)))
+        run(filename, tb=traceback)
+
+    return excepthook
 
 
 @contextmanager